when collating arguments, lazy-get implicits by type
this corrects a subtle bug when collating inherited arguments. previously, it would use container->get(), which would try to instantiate the inherited argument right then; if the inherited argument was not defined, this would cause an error, even if the child argument was defined. using lazy-get defers instantation of the argument until the moment it is actually used.